feat(login): fixed login form, token, user handling

+++ client/src/context/AuthContext/AuthContext.tsx
@@ -1,41 +1,58 @@
-import { createContext, Dispatch, SetStateAction, useCallback, useContext, useState } from "react";
+import { addJWTHeader, removeJWTHeader } from "$/config/backendClient";
+import {
+  createContext,
+  Dispatch,
+  SetStateAction,
+  useCallback,
+  useContext,
+  useEffect,
+  useState,
+} from "react";
 import JwtManager from "../../utils/JwtManager";
-
 
 interface AuthContext {
   user: string | null;
+  setToken: Dispatch<string | null>;
   setUser: Dispatch<SetStateAction<string | null>>;
   signOut: () => void;
 }
 
 const initalState: AuthContext = {
   setUser: (user) => user,
-  signOut: () => { },
-  user: null
-}
+  setToken: () => {},
+  signOut: () => {},
+  user: null,
+};
 
 const Context = createContext<AuthContext>(initalState);
 
-
 const AuthContextProvider: React.FC = ({ children }) => {
+  const [user, setUser] = useState<string | null>(initalState.user);
 
-  const [user, setUser] = useState<string | null>(initalState.user)
+  const [token, setToken] = useState(JwtManager.get());
 
+  useEffect(() => {
+    if (token) {
+      JwtManager.set(token);
+      addJWTHeader(token);
+    } else {
+      JwtManager.delete();
+      removeJWTHeader();
+      setUser(null);
+    }
+  }, [token]);
 
-  const signOut = useCallback(() => {
-    setUser(null);
-    JwtManager.delete();
-  }, []);
+  const signOut = useCallback(() => setToken(null), [setToken]);
 
-  return <Context.Provider value={{ user, setUser, signOut }}>
-    {children}
-  </Context.Provider>
-
-
-}
+  return (
+    <Context.Provider value={{ user, setUser, signOut, setToken }}>
+      {children}
+    </Context.Provider>
+  );
+};
 
 export default AuthContextProvider;
 
 const useAuth = () => useContext(Context);
 
-export { useAuth }
+export { useAuth };
